+++ examples/svelte-design-system/src/app.d.ts
@@ -1,0 +1,19 @@
+/// <reference types="@sveltejs/kit" />
+
+// See https://kit.svelte.dev/docs/types#app
+// for information about these interfaces
+declare namespace App {
+	interface Locals {
+		userid: string;
+	}
+
+	// interface Platform {}
+
+	// interface Session {}
+
+	// interface Stuff {}
+}
+
+declare module '@builder.io/sdk-svelte';
+
+declare module 'svelte-parallax';