+++ packages/react/src/components/builder-component.component.tsx
@@ -28,8 +28,8 @@ import {
 import { Url } from 'url';
 import { debounceNextTick } from '../functions/debonce-next-tick';
 import { throttle } from '../functions/throttle';
-import { safeDynamicRequire } from '../functions/safe-dynamic-require';
 import { BuilderMetaContext } from '../store/builder-meta';
+import { tryEval } from '../functions/try-eval';
 
 function pick<T, K extends keyof T>(obj: T, ...keys: K[]): Pick<T, K> {
   const ret: any = {};
@@ -73,11 +73,11 @@ const size = (thing: object) => Object.keys(thing).length;
 
 function debounce(func: Function, wait: number, immediate = false) {
   let timeout: any;
-  return function (this: any) {
+  return function(this: any) {
     const context = this;
     const args = arguments;
     clearTimeout(timeout);
-    timeout = setTimeout(function () {
+    timeout = setTimeout(function() {
       timeout = null;
       if (!immediate) func.apply(context, args);
     }, wait);
@@ -308,84 +308,6 @@ export interface BuilderComponentState {
   context: any;
   key: number;
 }
-
-const tryEval = (str?: string, data: any = {}, errors?: Error[]): any => {
-  const value = str;
-  if (!(typeof value === 'string' && value.trim())) {
-    return;
-  }
-  const useReturn = !(value.includes(';') || value.includes(' return '));
-  let fn: Function = () => {
-    /* Intentionally empty */
-  };
-  try {
-    if (Builder.isBrowser) {
-      // tslint:disable-next-line:no-function-constructor-with-string-args
-      // TODO: VM in node......
-      fn = new Function(
-        'state',
-        // TODO: remove the with () {} - make a page v3 that doesn't use this
-        `var rootState = state;
-        if (typeof Proxy !== 'undefined') {
-          rootState = new Proxy(rootState, {
-            set: function () {
-              return false;
-            },
-            get: function (target, key) {
-              if (key === 'state') {
-                return state;
-              }
-              return target[key]
-            }
-          });
-        }
-        with (rootState) {
-          ${useReturn ? `return (${str});` : str};
-        }`
-      );
-    }
-  } catch (error) {
-    if (Builder.isBrowser) {
-      console.warn('Could not compile javascript', error);
-    } else {
-      // Add to req.options.errors to return to client
-    }
-  }
-  try {
-    if (Builder.isBrowser) {
-      return fn(data || {});
-    } else {
-      // Below is a hack to get certain code to *only* load in the server build, to not screw with
-      // browser bundler's like rollup and webpack. Our rollup plugin strips these comments only
-      // for the server build
-      // tslint:disable:comment-format
-      const { VM } = safeDynamicRequire('vm2');
-      return new VM({
-        sandbox: {
-          ...data,
-          ...{ state: data },
-        },
-        // TODO: convert reutrn to module.exports on server
-      }).run(value.replace(/(^|;)return /, '$1'));
-      // tslint:enable:comment-format
-    }
-  } catch (error: any) {
-    if (errors) {
-      errors.push(error);
-    }
-
-    if (Builder.isBrowser) {
-      console.warn('Builder custom code error:', error.message, 'in', str, error.stack);
-    } else {
-      if (process.env.DEBUG) {
-        console.debug('Builder custom code error:', error.message, 'in', str, error.stack);
-      }
-      // Add to req.options.errors to return to client
-    }
-  }
-
-  return;
-};
 
 function searchToObject(location: Location | Url) {
   const pairs = (location.search || '').substring(1).split('&');
@@ -1110,7 +1032,7 @@ export class BuilderComponent extends React.Component<
                           const useBuilderState = (initialState: any) => {
                             const [, setTick] = React.useState(0);
                             const [state] = React.useState(() =>
-                              onChange(initialState, function () {
+                              onChange(initialState, function() {
                                 setTick(tick => tick + 1);
                               })
                             );
@@ -1417,7 +1339,7 @@ export class BuilderComponent extends React.Component<
 
           // TODO: allow exports = { } syntax?
           // TODO: do something with reuslt like view - methods, computed, actions, properties, template, etc etc
-        } catch (error: any) {
+        } catch (error) {
           if (Builder.isBrowser) {
             console.warn(
               'Builder custom code error:',
@@ -1481,14 +1403,15 @@ export class BuilderComponent extends React.Component<
                 }
 
                 // TODO: fix this
-                const newSubscription = (this.httpSubscriptionPerKey[key] =
-                  this.onStateChange.subscribe(() => {
-                    const newUrl = this.evalExpression(url);
-                    if (newUrl !== finalUrl) {
-                      this.handleRequest(key, newUrl);
-                      this.lastHttpRequests[key] = newUrl;
-                    }
-                  }));
+                const newSubscription = (this.httpSubscriptionPerKey[
+                  key
+                ] = this.onStateChange.subscribe(() => {
+                  const newUrl = this.evalExpression(url);
+                  if (newUrl !== finalUrl) {
+                    this.handleRequest(key, newUrl);
+                    this.lastHttpRequests[key] = newUrl;
+                  }
+                }));
                 this.subscriptions.add(newSubscription);
               }
             } else {